refactor: Remove adapter and dataset
