+++ ImageFeed/Screens/ImagesList/ImagesListViewController.swift
@@ -9,8 +9,10 @@ import UIKit
 
 final class ImagesListViewController: UIViewController {
         
+    private var pictures: [Picture] = []
+    private var didAnimateCells: [IndexPath: Bool] = [:]
+    
     private lazy var tableView = UITableView()
-    private let adapter: IImagesListTableViewAdapter
     
     private lazy var refreshControl = UIRefreshControl()
     private var hasRefreshed = false {
@@ -19,8 +21,9 @@ final class ImagesListViewController: UIViewController {
         }
     }
     
-    init(adapter: IImagesListTableViewAdapter) {
-        self.adapter = adapter
+    var onSelect: ((Picture) -> Void)?
+    
+    init() {
         super.init(nibName: nil, bundle: nil)
     }
     
@@ -39,9 +42,18 @@ final class ImagesListViewController: UIViewController {
 
 private extension ImagesListViewController {
     private func setup() {
-        adapter.setupTableView(tableView)
+        pictures = MockProvider.pictures
+        setupTableView()
         
         refreshControl.addTarget(self, action: #selector(refreshContent), for: .valueChanged)
+    }
+    
+    func setupTableView() {
+        tableView.register(models: [PictureViewModel.self])
+        
+        tableView.dataSource = self
+        tableView.delegate = self
+        
         tableView.refreshControl = refreshControl
     }
     
@@ -72,12 +84,65 @@ private extension ImagesListViewController {
 // MARK: - Actions
 private extension ImagesListViewController {
     @objc func refreshContent() {
-        adapter.shufflePictures()
+        pictures.shuffle()
+        didAnimateCells = [:]
         hasRefreshed.toggle()
     }
     
     func reloadView() {
         tableView.refreshControl?.endRefreshing()
-        adapter.reloadTableView(tableView)
+        DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) { [weak self] in
+            guard let self = self else { return }
+            self.tableView.performBatchUpdates {
+                self.tableView.reloadSections([0], with: .automatic)
+            }
+        }
+    }
+}
+
+// MARK: - UITableViewDataSource
+extension ImagesListViewController: UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        pictures.count
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        guard let picture = pictures[safe: indexPath.row] else { return UITableViewCell() }
+        let model = PictureViewModel(from: picture) { [weak self] in
+            self?.pictures[indexPath.row].isFavorite.toggle()
+        }
+        
+        return tableView.dequeueReusableCell(withModel: model, for: indexPath)
+    }
+}
+
+// MARK: - UITableViewDelegate
+extension ImagesListViewController: UITableViewDelegate {
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        tableView.deselectRow(at: indexPath, animated: true)
+        guard
+            let picture = pictures[safe: indexPath.row],
+            let onSelect = onSelect
+        else {
+            return
+        }
+        onSelect(picture)
+    }
+    
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        guard let picture = pictures[safe: indexPath.row] else { return 0 }
+        return Theme.size(kind: .cellHeight(image: UIImage(named: picture.image)))
+    }
+    
+    func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
+        guard didAnimateCells[indexPath] == nil else { return }
+        didAnimateCells[indexPath] = true
+        
+        cell.transform3DMakeRotation(
+            degree: 90,
+            x: 0, y: 1, z: 0,
+            duration: 0.85,
+            delay: 0.1
+        )
     }
 }
